Rename project to FuckBF
