+++ .github/workflows/release.yml
@@ -127,15 +127,15 @@ jobs:
 
       - name: Rename binary (Linux & macOS)
         if: matrix.target != 'x86_64-pc-windows-gnu' && matrix.target != 'i686-pc-windows-gnu'
-        run: mv target/${{ matrix.target }}/release/fbf target/${{ matrix.target }}/release/fbf-${{ matrix.name }}
+        run: mv target/${{ matrix.target }}/release/fuckbf target/${{ matrix.target }}/release/fuckbf-${{ matrix.name }}
 
       - name: Rename binary (Windows)
         if: matrix.target == 'x86_64-pc-windows-gnu' || matrix.target == 'i686-pc-windows-gnu'
-        run: mv target/${{ matrix.target }}/release/fbf.exe target/${{ matrix.target }}/release/fbf-${{ matrix.name }}.exe
+        run: mv target/${{ matrix.target }}/release/fuckbf.exe target/${{ matrix.target }}/release/fuckbf-${{ matrix.name }}.exe
 
 
       - name: Upload release
         uses: softprops/action-gh-release@v1
         with:
-          files: target/${{ matrix.target }}/release/fbf-${{ matrix.name }}*
+          files: target/${{ matrix.target }}/release/fuckbf-${{ matrix.name }}*
           token: ${{ secrets.PAT_GITHUB }}
