+++ .gitignore
@@ -4,6 +4,7 @@
 # Test
 test.bf
 test.fbf
+test.fuckbf
 
 # IDEs
 .idea/
