+++ build.rs
@@ -1,7 +1,7 @@
 fn main() {
     // Save target for update command
     println!(
-        "cargo:rustc-env=FUCKBRAINFUCK_TARGET_OS={}",
+        "cargo:rustc-env=FUCKBF_TARGET_OS={}",
         &std::env::var("TARGET").unwrap()
     );
 }
